+++ src/main_loop/proof_upgrader.rs
@@ -851,11 +851,11 @@ mod tests {
     use crate::models::state::tx_creation_config::TxCreationConfig;
     use crate::models::state::wallet::address::generation_address::GenerationReceivingAddress;
     use crate::models::state::wallet::transaction_output::TxOutput;
-    use crate::tests::shared::fake_block_successor_with_merged_tx;
-    use crate::tests::shared::get_test_genesis_setup;
-    use crate::tests::shared::invalid_empty_block_with_timestamp;
-    use crate::tests::shared::mock_genesis_global_state;
-    use crate::tests::shared::state_with_premine_and_self_mined_blocks;
+    use crate::tests::shared::blocks::fake_block_successor_with_merged_tx;
+    use crate::tests::shared::blocks::invalid_empty_block_with_timestamp;
+    use crate::tests::shared::globalstate::get_test_genesis_setup;
+    use crate::tests::shared::globalstate::mock_genesis_global_state;
+    use crate::tests::shared::globalstate::state_with_premine_and_self_mined_blocks;
     use crate::tests::shared_tokio_runtime;
     use crate::PEER_CHANNEL_CAPACITY;
