+++ src/models/peer.rs
@@ -915,7 +915,7 @@ mod tests {
     use super::*;
     use crate::models::blockchain::block::block_header::HeaderToBlockHashWitness;
     use crate::models::blockchain::block::Block;
-    use crate::tests::shared::fake_valid_sequence_of_blocks_for_tests;
+    use crate::tests::shared::blocks::fake_valid_sequence_of_blocks_for_tests;
     use crate::tests::shared_tokio_runtime;
 
     impl PeerStanding {
