+++ src/models/state/wallet/mod.rs
@@ -63,11 +63,11 @@ mod tests {
     use crate::models::state::wallet::transaction_output::TxOutputList;
     use crate::models::state::wallet::wallet_entropy::WalletEntropy;
     use crate::models::state::GlobalStateLock;
-    use crate::tests::shared::invalid_block_with_transaction;
-    use crate::tests::shared::make_mock_block;
-    use crate::tests::shared::make_mock_transaction_with_mutator_set_hash;
-    use crate::tests::shared::mock_genesis_global_state;
+    use crate::tests::shared::blocks::invalid_block_with_transaction;
+    use crate::tests::shared::blocks::make_mock_block;
+    use crate::tests::shared::globalstate::mock_genesis_global_state;
     use crate::tests::shared::mock_genesis_wallet_state;
+    use crate::tests::shared::mock_tx::make_mock_transaction_with_mutator_set_hash;
     use crate::tests::shared_tokio_runtime;
     use crate::triton_vm_job_queue::TritonVmJobPriority;
     use crate::triton_vm_job_queue::TritonVmJobQueue;
